requirements.txt
absl-py==1.1.0
aiohttp==3.8.1
aiosignal==1.2.0
anytree==2.8.0
astunparse==1.6.3
async-timeout==4.0.2
asynctest==0.13.0
attrs==21.4.0
blis==0.7.8
cachetools==5.2.0
catalogue==2.0.7
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
cymem==2.0.6
datasets==2.3.2
dill==0.3.5.1
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.3.0/en_core_web_sm-3.3.0-py3-none-any.whl
filelock==3.7.1
flatbuffers==1.12
frozenlist==1.3.0
fsspec==2022.5.0
gast==0.4.0
google-auth==2.8.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.47.0
h5py==3.7.0
huggingface-hub==0.8.1
idna==3.3
importlib-metadata==4.11.4
Jinja2==3.1.2
joblib==1.1.0
keras==2.9.0
Keras-Preprocessing==1.1.2
langcodes==3.3.0
libclang==14.0.1
Markdown==3.3.7
MarkupSafe==2.1.1
multidict==6.0.2
multiprocess==0.70.13
murmurhash==1.0.7
nltk==3.7
numpy==1.21.6
oauthlib==3.2.0
OpenHowNet==2.0
opt-einsum==3.3.0
packaging==21.3
pandas==1.3.5
pathy==0.6.1
Pillow==9.2.0
preshed==3.0.6
protobuf==3.19.4
pyarrow==8.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.8.2
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
regex==2022.6.2
requests==2.28.0
requests-oauthlib==1.3.1
responses==0.18.0
rsa==4.8
scikit-learn==1.0.2
scipy==1.7.3
sentence-transformers==2.2.2
sentencepiece==0.1.96
six==1.16.0
smart-open==5.2.1
spacy==3.3.1
spacy-legacy==3.0.9
spacy-loggers==1.0.2
srsly==2.4.3
stanfordnlp==0.2.0
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.9.1
tensorflow-estimator==2.9.0
tensorflow-hub==0.12.0
tensorflow-io-gcs-filesystem==0.26.0
tensorflow-text==2.9.0
termcolor==1.1.0
thinc==8.0.17
threadpoolctl==3.1.0
tokenizers==0.12.1
torch==1.12.0
torchvision==0.13.0
tqdm==4.64.0
transformers==4.20.1
typer==0.4.1
typing_extensions==4.1.1
urllib3==1.26.9
wasabi==0.9.1
Werkzeug==2.1.2
wrapt==1.14.1
xxhash==3.0.0
yarl==1.7.2
zipp==3.8.0
